refactor(frontend): Move request/response model to editor module
