+++ frontend/src/app/editor/diagram-view/diagram-view.component.ts
@@ -1,7 +1,6 @@
 import {Component, Input} from '@angular/core';
 import {environment} from '../../../environments/environment';
-import Diagram from '../../model/codegen/diagram';
-import Response from '../../model/codegen/response';
+import {Diagram, Response} from '../model/response';
 
 @Component({
   selector: 'app-diagram-view',
