--- frontend/src/app/model/codegen/diagram.ts
@@ -1,5 +1,0 @@
-export default class Diagram {
-  path: string;
-  name: string;
-  content: string;
-}
