--- frontend/src/app/model/codegen/test-method.ts
@@ -1,5 +1,0 @@
-export default class TestMethod {
-  className: string;
-  name: string;
-  body: string;
-}
