+++ frontend/src/app/scenario-editor.service.ts
@@ -1,27 +1,10 @@
-import {HttpClient} from '@angular/common/http';
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs';
-
-import {environment} from '../environments/environment';
-import Request from './model/codegen/request';
-import Response from './model/codegen/response';
-import {PrivacyService} from './privacy.service';
 
 
 @Injectable({
   providedIn: 'root',
 })
 export class ScenarioEditorService {
-  constructor(
-    private privacyService: PrivacyService,
-    private http: HttpClient,
-  ) {
-  }
-
-  submit(codeGenRequest: Request): Observable<Response> {
-    return this.http.post<Response>(environment.apiURL + '/runcodegen', codeGenRequest);
-  }
-
   foldInternalCalls(packageName: string, outputLines: string[]): string[] {
     packageName = packageName.replace('/', '.');
     const packageNamePrefix = `\tat ${packageName}.`;
